Channel-specific artifact handling
Artifacts may now be provided separately for each channel (as a cell array of artifacts). All other ways to define artifacts will be turned into this form.

Importantly, artifacts are not anymore marked in scoring_fine but handled by a local function isArt(samples, arts) which returns 0s for non-artifactual sampled and 1s for artifactual ones. As a result, the spectrum is now calculated on data *without* artifact removal.

Also removes output of fine_scoring with artifacts because for many channels, that can be quite large.